+++ src/bare-metal/linux-x86/hw_specific.cpp
@@ -28,7 +28,7 @@ namespace RODOS {
  */
 
 extern void sigtermHandler(int sig);
-extern long myNodeNr;
+extern int32_t myNodeNr;
 
 //_____________________________________________________
 void hwInit() {
